bookmark/requirements.txt
django-tastypie
django
requests
